requirements.txt
Flask
google-genai
python-dotenv
PyMuPDF
geopy
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
plotly
